+++ src/main/java/io/milvus/client/GetTableRowCountResponse.java
@@ -40,6 +40,10 @@ public class GetTableRowCountResponse {
     return response;
   }
 
+  public boolean ok() {
+    return response.ok();
+  }
+
   @Override
   public String toString() {
     return String.format(