+++ src/main/java/io/milvus/client/InsertResponse.java
@@ -41,6 +41,10 @@ public class InsertResponse {
     return response;
   }
 
+  public boolean ok() {
+    return response.ok();
+  }
+
   @Override
   public String toString() {
     return String.format(