+++ src/main/java/io/milvus/client/ShowTablesResponse.java
@@ -42,6 +42,10 @@ public class ShowTablesResponse {
     return response;
   }
 
+  public boolean ok() {
+    return response.ok();
+  }
+
   @Override
   public String toString() {
     return String.format(